cambio de tipos de datos

+++ ServidorJava/GameRules.java
@@ -1,22 +1,20 @@
-
-
 public final class GameRules {
-    public static final int MAX_LIANAS = 6;
-    public static final int HEIGHT_MAX = 12;
+    public static final Integer MAX_LIANAS = Integer.valueOf(6);
+    public static final Integer HEIGHT_MAX = Integer.valueOf(12);
 
     // Límites del mundo en píxeles
-    public static final float MIN_X = 20.0f;
-    public static final float MAX_X = 780.0f;
-    public static final float MIN_Y = 0.0f;
-    public static final float MAX_Y = 600.0f;
+    public static final Float MIN_X = Float.valueOf(20.0f);
+    public static final Float MAX_X = Float.valueOf(780.0f);
+    public static final Float MIN_Y = Float.valueOf(0.0f);
+    public static final Float MAX_Y = Float.valueOf(600.0f);
 
     // Física
-    public static final float GRAVITY     = 500.0f;
-    public static final float MOVE_SPEED  = 50.0f;
-    public static final float JUMP_FORCE  = 250.0f;
-    public static final float CLIMB_SPEED = 75.0f;
+    public static final Float GRAVITY     = Float.valueOf(500.0f);
+    public static final Float MOVE_SPEED  = Float.valueOf(50.0f);
+    public static final Float JUMP_FORCE  = Float.valueOf(250.0f);
+    public static final Float CLIMB_SPEED = Float.valueOf(75.0f);
 
-    public static final float MAX_FALL_SPEED = 400.0f;  // Velocidad máxima de caída
+    public static final Float MAX_FALL_SPEED = Float.valueOf(400.0f);  // Velocidad máxima de caída
 
     private GameRules() {}
 }
